refactor: extract sleepWithAbort as reusable utility function
- Add sleepWithAbort() to src/utils/helpers.ts for abort-aware delays
- Remove private waitWithAbort() method from ModelHandlerOpenAIResponse
- Uses AbortSignal.timeout() when available for better performance
- Throws DOMException with 'AbortError' name on cancellation
- Reusable by any code needing cancellable delays (polling, retries, etc.)
